linera_client

Module benchmark

Tooling for running throughput benchmarks against a network.

Structs

Benchmark

Driver for running benchmarks against a network.

BenchmarkConfig

Configuration listing the chains to use for a benchmark.

FungibleTransferGenerator

Generates fungible token transfer operations between chains.

NativeFungibleTransferGenerator

Generates native fungible token transfer operations between chains.

Enums

BenchmarkError

An error that can occur while running a benchmark.

Traits

OperationGenerator

Trait for generating benchmark operations.

Functions

fungible_transfer

Builds a fungible token transfer operation for the given application.