Struct linera_client::client_context::ClientContext

pub struct ClientContext<Storage, W>
where
    Storage: Storage,
{
    pub wallet: WalletState<W>,
    pub client: Arc<Client<NodeProvider, Storage>>,
    pub send_timeout: Duration,
    pub recv_timeout: Duration,
    pub retry_delay: Duration,
    pub max_retries: u32,
    pub chain_listeners: JoinSet,
    pub blanket_message_policy: BlanketMessagePolicy,
    pub restrict_chain_ids_to: Option<HashSet<ChainId>>,
}

Fields
wallet: WalletState<W>client: Arc<Client<NodeProvider, Storage>>send_timeout: Durationrecv_timeout: Durationretry_delay: Durationmax_retries: u32chain_listeners: JoinSetblanket_message_policy: BlanketMessagePolicyrestrict_chain_ids_to: Option<HashSet<ChainId>>
Implementations

impl<S, W> ClientContext<S, W>
where S: Storage + Clone + Send + Sync + 'static, W: Persist<Target = Wallet>,

pub fn wallet(&self) -> &Wallet

Returns a reference to the wallet.

pub fn wallet_mut(&mut self) -> &mut WalletState<W>

Returns the WalletState as a mutable reference.

pub async fn mutate_wallet<R: Send>( &mut self, mutation: impl FnOnce(&mut Wallet) -> R + Send, ) -> Result<R, Error>

pub fn new(storage: S, options: ClientOptions, wallet: W) -> Self

pub fn new_test_client_context(storage: S, wallet: W) -> Self

pub fn default_account(&self) -> Account

Retrieve the default account. Current this is the common account of the default chain.

pub fn default_chain(&self) -> ChainId

Retrieve the default chain.

pub fn make_node_provider(&self) -> NodeProvider

pub async fn save_wallet(&mut self) -> Result<(), Error>

pub async fn update_and_save_wallet( &mut self, client: &ChainClient<NodeProvider, S>, ) -> Result<(), Error>

pub async fn update_wallet_for_new_chain( &mut self, chain_id: ChainId, key_pair: Option<KeyPair>, timestamp: Timestamp, ) -> Result<(), Error>

Remembers the new chain and private key (if any) in the wallet.

pub async fn process_inbox( &mut self, chain_client: &ChainClient<NodeProvider, S>, ) -> Result<Vec<ConfirmedBlockCertificate>, Error>

pub async fn apply_client_command<E, F, Fut, T>( &mut self, client: &ChainClient<NodeProvider, S>, f: F, ) -> Result<T, Error>
where F: FnMut(&ChainClient<NodeProvider, S>) -> Fut, Fut: Future<Output = Result<ClientOutcome<T>, E>>, Error: From<E>,

Applies the given function to the chain client.

Updates the wallet regardless of the outcome. As long as the function returns a round timeout, it will wait and retry.

pub async fn change_ownership( &mut self, chain_id: Option<ChainId>, ownership_config: ChainOwnershipConfig, ) -> Result<(), Error>

impl<S, W> ClientContext<S, W>
where S: Storage + Clone + Send + Sync + 'static, W: Persist<Target = Wallet>,

pub async fn publish_bytecode( &mut self, chain_client: &ChainClient<NodeProvider, S>, contract: PathBuf, service: PathBuf, ) -> Result<BytecodeId, Error>

pub async fn publish_data_blob( &mut self, chain_client: &ChainClient<NodeProvider, S>, blob_path: PathBuf, ) -> Result<CryptoHash, Error>

pub async fn read_data_blob( &mut self, chain_client: &ChainClient<NodeProvider, S>, hash: CryptoHash, ) -> Result<(), Error>

impl<S, W> ClientContext<S, W>
where S: Storage + Clone + Send + Sync + 'static, W: Persist<Target = Wallet>,

pub async fn process_inboxes_and_force_validator_updates(&mut self)

pub async fn make_benchmark_chains( &mut self, num_chains: usize, balance: Amount, ) -> Result<HashMap<ChainId, KeyPair>, Error>

Creates chains if necessary, and returns a map of exactly num_chains chain IDs with key pairs.

pub async fn supply_fungible_tokens( &mut self, key_pairs: &HashMap<ChainId, KeyPair>, application_id: ApplicationId, max_in_flight: usize, ) -> Result<(), Error>

Creates chains if necessary, and returns a map of exactly num_chains chain IDs with key pairs.

pub fn make_benchmark_block_proposals( &mut self, key_pairs: &HashMap<ChainId, KeyPair>, transactions_per_block: usize, fungible_application_id: Option<ApplicationId>, ) -> Vec<RpcMessage>

Makes one block proposal per chain, up to num_chains blocks.

pub fn make_benchmark_certificates_from_votes<T>( &self, votes: Vec<Vote<T>>, ) -> Vec<GenericCertificate<T>>
where T: Debug + CertificateValue,

Tries to aggregate votes into certificates.

pub async fn mass_broadcast( &self, phase: &'static str, max_in_flight: usize, proposals: Vec<RpcMessage>, ) -> Vec<RpcMessage>

Broadcasts a bulk of blocks to each validator.

pub async fn update_wallet_from_certificates( &mut self, certificates: Vec<ConfirmedBlockCertificate>, )

pub async fn stage_block_execution( &self, block: Block, ) -> Result<ExecutedBlock, Error>

Stages the execution of a block proposal.