Enum linera_client::client_options::ClientCommand

pub enum ClientCommand {

    Transfer {
        sender: Account,
        recipient: Account,
        amount: Amount,
    },
    OpenChain {
        chain_id: Option<ChainId>,
        owner: Option<Owner>,
        balance: Amount,
    },
    OpenMultiOwnerChain {
        chain_id: Option<ChainId>,
        ownership_config: ChainOwnershipConfig,
        application_permissions_config: ApplicationPermissionsConfig,
        balance: Amount,
    },
    ChangeOwnership {
        chain_id: Option<ChainId>,
        ownership_config: ChainOwnershipConfig,
    },
    ChangeApplicationPermissions {
        chain_id: Option<ChainId>,
        application_permissions_config: ApplicationPermissionsConfig,
    },
    CloseChain {
        chain_id: ChainId,
    },
    LocalBalance {
        account: Option<Account>,
    },
    QueryBalance {
        account: Option<Account>,
    },
    SyncBalance {
        account: Option<Account>,
    },
    Sync {
        chain_id: Option<ChainId>,
    },
    ProcessInbox {
        chain_id: Option<ChainId>,
    },
    QueryValidator {
        address: String,
        chain_id: Option<ChainId>,
        public_key: Option<ValidatorPublicKey>,
    },
    QueryValidators {
        chain_id: Option<ChainId>,
    },
    SyncValidator {
        address: String,
        chains: Vec<ChainId>,
    },
    SetValidator {
        public_key: ValidatorPublicKey,
        account_key: AccountPublicKey,
        address: String,
        votes: u64,
        skip_online_check: bool,
    },
    RemoveValidator {
        public_key: ValidatorPublicKey,
    },
    FinalizeCommittee,
    ResourceControlPolicy {

        block: Option<Amount>,
        fuel_unit: Option<Amount>,
        read_operation: Option<Amount>,
        write_operation: Option<Amount>,
        byte_read: Option<Amount>,
        byte_written: Option<Amount>,
        byte_stored: Option<Amount>,
        operation: Option<Amount>,
        operation_byte: Option<Amount>,
        message: Option<Amount>,
        message_byte: Option<Amount>,
        maximum_fuel_per_block: Option<u64>,
        maximum_executed_block_size: Option<u64>,
        maximum_blob_size: Option<u64>,
        maximum_published_blobs: Option<u64>,
        maximum_bytecode_size: Option<u64>,
        maximum_block_proposal_size: Option<u64>,
        maximum_bytes_read_per_block: Option<u64>,
        maximum_bytes_written_per_block: Option<u64>,
    
},
    Benchmark {
        num_chains: usize,
        tokens_per_chain: Amount,
        transactions_per_block: usize,
        fungible_application_id: Option<ApplicationId>,
        bps: Option<usize>,
        close_chains: bool,
        health_check_endpoints: Option<String>,
        wrap_up_max_in_flight: usize,
    },
    CreateGenesisConfig {

        committee_config_path: PathBuf,
        genesis_config_path: PathBuf,
        admin_root: u32,
        initial_funding: Amount,
        start_timestamp: Option<DateTime<Utc>>,
        num_other_initial_chains: u32,
        policy_config: ResourceControlPolicyConfig,
        block_price: Option<Amount>,
        fuel_unit_price: Option<Amount>,
        read_operation_price: Option<Amount>,
        write_operation_price: Option<Amount>,
        byte_read_price: Option<Amount>,
        byte_written_price: Option<Amount>,
        byte_stored_price: Option<Amount>,
        operation_price: Option<Amount>,
        operation_byte_price: Option<Amount>,
        message_price: Option<Amount>,
        message_byte_price: Option<Amount>,
        maximum_fuel_per_block: Option<u64>,
        maximum_executed_block_size: Option<u64>,
        maximum_bytecode_size: Option<u64>,
        maximum_blob_size: Option<u64>,
        maximum_published_blobs: Option<u64>,
        maximum_block_proposal_size: Option<u64>,
        maximum_bytes_read_per_block: Option<u64>,
        maximum_bytes_written_per_block: Option<u64>,
        http_allow_list: Option<Vec<String>>,
        testing_prng_seed: Option<u64>,
        network_name: Option<String>,
    
},
    Watch {
        chain_id: Option<ChainId>,
        raw: bool,
    },
    Service {
        config: ChainListenerConfig,
        port: NonZeroU16,
    },
    Faucet {
        chain_id: Option<ChainId>,
        port: NonZeroU16,
        amount: Amount,
        limit_rate_until: Option<DateTime<Utc>>,
        config: ChainListenerConfig,
    },
    PublishModule {
        contract: PathBuf,
        service: PathBuf,
        vm_runtime: VmRuntime,
        publisher: Option<ChainId>,
    },
    PublishDataBlob {
        blob_path: PathBuf,
        publisher: Option<ChainId>,
    },
    ReadDataBlob {
        hash: CryptoHash,
        reader: Option<ChainId>,
    },
    CreateApplication {
        module_id: ModuleId,
        creator: Option<ChainId>,
        json_parameters: Option<String>,
        json_parameters_path: Option<PathBuf>,
        json_argument: Option<String>,
        json_argument_path: Option<PathBuf>,
        required_application_ids: Option<Vec<UserApplicationId>>,
    },
    PublishAndCreate {
        contract: PathBuf,
        service: PathBuf,
        vm_runtime: VmRuntime,
        publisher: Option<ChainId>,
        json_parameters: Option<String>,
        json_parameters_path: Option<PathBuf>,
        json_argument: Option<String>,
        json_argument_path: Option<PathBuf>,
        required_application_ids: Option<Vec<UserApplicationId>>,
    },
    Keygen,
    Assign {
        owner: Owner,
        message_id: MessageId,
    },
    RetryPendingBlock {
        chain_id: Option<ChainId>,
    },
    Wallet(WalletCommand),
    Project(ProjectCommand),
    Net(NetCommand),
    Storage(DatabaseToolCommand),
    HelpMarkdown,
    ExtractScriptFromMarkdown {
        path: PathBuf,
        pause_after_linera_service: Duration,
        pause_after_gql_mutations: Duration,
    },

}

Variants

Transfer

Transfer funds

Fields

sender: Account

Sending chain ID (must be one of our chains)

recipient: Account

Recipient account

amount: Amount

Amount to transfer

OpenChain

Open (i.e. activate) a new chain deriving the UID from an existing one.

Fields

chain_id: Option<ChainId>

Chain ID (must be one of our chains).

owner: Option<Owner>

The new owner (otherwise create a key pair and remember it)

balance: Amount

The initial balance of the new chain. This is subtracted from the parent chain’s balance.

OpenMultiOwnerChain

Open (i.e. activate) a new multi-owner chain deriving the UID from an existing one.

Fields

chain_id: Option<ChainId>

Chain ID (must be one of our chains).

ownership_config: ChainOwnershipConfig

application_permissions_config: ApplicationPermissionsConfig

balance: Amount

The initial balance of the new chain. This is subtracted from the parent chain’s balance.

ChangeOwnership

Change who owns the chain, and how the owners work together proposing blocks.

Specify the complete set of new owners, by public key. Existing owners that are not included will be removed.

Fields

chain_id: Option<ChainId>

The ID of the chain whose owners will be changed.

ownership_config: ChainOwnershipConfig

ChangeApplicationPermissions

Changes the application permissions configuration.

Fields

chain_id: Option<ChainId>

The ID of the chain to which the new permissions will be applied.

application_permissions_config: ApplicationPermissionsConfig

CloseChain

Close an existing chain.

A closed chain cannot execute operations or accept messages anymore. It can still reject incoming messages, so they bounce back to the sender.

Fields

chain_id: ChainId

Chain ID (must be one of our chains)

LocalBalance

Read the current native-token balance of the given account directly from the local state.

NOTE: The local balance does not reflect messages that are waiting to be picked in the local inbox, or that have not been synchronized from validators yet. Use linera sync then either linera query-balance or linera process-inbox && linera local-balance for a consolidated balance.

Fields

account: Option<Account>

The account to read, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

QueryBalance

Simulate the execution of one block made of pending messages from the local inbox, then read the native-token balance of the account from the local state.

NOTE: The balance does not reflect messages that have not been synchronized from validators yet. Call linera sync first to do so.

Fields

account: Option<Account>

The account to query, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

SyncBalance

(DEPRECATED) Synchronize the local state of the chain with a quorum validators, then query the local balance.

This command is deprecated. Use linera sync && linera query-balance instead.

Fields

account: Option<Account>

The account to query, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

Sync

Synchronize the local state of the chain with a quorum validators.

Fields

chain_id: Option<ChainId>

The chain to synchronize with validators. If omitted, synchronizes the default chain of the wallet.

ProcessInbox

Process all pending incoming messages from the inbox of the given chain by creating as many blocks as needed to execute all (non-failing) messages. Failing messages will be marked as rejected and may bounce to their sender depending on their configuration.

Fields

chain_id: Option<ChainId>

The chain to process. If omitted, uses the default chain of the wallet.

QueryValidator

Show the version and genesis config hash of a new validator, and print a warning if it is incompatible. Also print some information about the given chain while we are at it.

Fields

address: String

The new validator’s address.

chain_id: Option<ChainId>

The chain to query. If omitted, query the default chain of the wallet.

public_key: Option<ValidatorPublicKey>

The public key of the validator. If given, the signature of the chain query info will be checked.

QueryValidators

Show the current set of validators for a chain. Also print some information about the given chain while we are at it.

Fields

chain_id: Option<ChainId>

The chain to query. If omitted, query the default chain of the wallet.

SyncValidator

Synchronizes a validator with the local state of chains.

Fields

address: String

The public address of the validator to synchronize.

chains: Vec<ChainId>

The chains to synchronize, or the default chain if empty.

SetValidator

Add or modify a validator (admin only)

Fields

public_key: ValidatorPublicKey

The public key of the validator.

account_key: AccountPublicKey

The public key of the account controlled by the validator.

address: String

Network address

votes: u64

Voting power

skip_online_check: bool

Skip the version and genesis config checks.

RemoveValidator

Remove a validator (admin only)

Fields

public_key: ValidatorPublicKey

The public key of the validator.

FinalizeCommittee

Deprecates all committees except the last one.

ResourceControlPolicy

View or update the resource control policy

Fields

block: Option<Amount>

Set the base price for creating a block.

fuel_unit: Option<Amount>

Set the price per unit of fuel.

read_operation: Option<Amount>

Set the price per read operation.

write_operation: Option<Amount>

Set the price per write operation.

byte_read: Option<Amount>

Set the price per byte read.

byte_written: Option<Amount>

Set the price per byte written.

byte_stored: Option<Amount>

Set the price per byte stored.

operation: Option<Amount>

Set the base price of sending an operation from a block..

operation_byte: Option<Amount>

Set the additional price for each byte in the argument of a user operation.

message: Option<Amount>

Set the base price of sending a message from a block..

message_byte: Option<Amount>

Set the additional price for each byte in the argument of a user message.

maximum_fuel_per_block: Option<u64>

Set the maximum amount of fuel per block.

maximum_executed_block_size: Option<u64>

Set the maximum size of an executed block, in bytes.

maximum_blob_size: Option<u64>

Set the maximum size of data blobs, compressed bytecode and other binary blobs, in bytes.

maximum_published_blobs: Option<u64>

Set the maximum number of published blobs per block.

maximum_bytecode_size: Option<u64>

Set the maximum size of decompressed contract or service bytecode, in bytes.

maximum_block_proposal_size: Option<u64>

Set the maximum size of a block proposal, in bytes.

maximum_bytes_read_per_block: Option<u64>

Set the maximum read data per block.

maximum_bytes_written_per_block: Option<u64>

Set the maximum write data per block.

Benchmark

Send one transfer per chain in bulk mode

Fields

num_chains: usize

How many chains to use for the benchmark

tokens_per_chain: Amount

How many tokens to assign to each newly created chain. These need to cover the transaction fees per chain for the benchmark.

transactions_per_block: usize

How many transactions to put in each block.

fungible_application_id: Option<ApplicationId>

The application ID of a fungible token on the wallet’s default chain. If none is specified, the benchmark uses the native token.

bps: Option<usize>

If provided, will be long running, and block proposals will be sent at the provided fixed BPS rate.

close_chains: bool

If provided, will close the chains after the benchmark is finished. Keep in mind that closing the chains might take a while, and will increase the validator latency while they’re being closed.

health_check_endpoints: Option<String>

A comma-separated list of host:port pairs to query for health metrics. If provided, the benchmark will check these endpoints for validator health and terminate if any validator is unhealthy. Example: “127.0.0.1:21100,validator-1.some-network.linera.net:21100”

wrap_up_max_in_flight: usize

The maximum number of in-flight requests to validators when wrapping up the benchmark. While wrapping up, this controls the concurrency level when processing inboxes and closing chains.

CreateGenesisConfig

Create genesis configuration for a Linera deployment. Create initial user chains and print information to be used for initialization of validator setup. This will also create an initial wallet for the owner of the initial “root” chains.

Fields

committee_config_path: PathBuf

Sets the file describing the public configurations of all validators

genesis_config_path: PathBuf

The output config path to be consumed by the server

admin_root: u32

Index of the admin chain in the genesis config

initial_funding: Amount

Known initial balance of the chain

start_timestamp: Option<DateTime<Utc>>

The start timestamp: no blocks can be created before this time.

num_other_initial_chains: u32

Number of initial (aka “root”) chains to create in addition to the admin chain.

policy_config: ResourceControlPolicyConfig

Configure the resource control policy (notably fees) according to pre-defined settings.

block_price: Option<Amount>

Set the base price for creating a block. (This will overwrite value from --policy-config)

fuel_unit_price: Option<Amount>

Set the price per unit of fuel. (This will overwrite value from --policy-config)

read_operation_price: Option<Amount>

Set the price per read operation. (This will overwrite value from --policy-config)

write_operation_price: Option<Amount>

Set the price per write operation. (This will overwrite value from --policy-config)

byte_read_price: Option<Amount>

Set the price per byte read. (This will overwrite value from --policy-config)

byte_written_price: Option<Amount>

Set the price per byte written. (This will overwrite value from --policy-config)

byte_stored_price: Option<Amount>

Set the price per byte stored. (This will overwrite value from --policy-config)

operation_price: Option<Amount>

Set the base price of sending an operation from a block.. (This will overwrite value from --policy-config)

operation_byte_price: Option<Amount>

Set the additional price for each byte in the argument of a user operation. (This will overwrite value from --policy-config)

message_price: Option<Amount>

Set the base price of sending a message from a block.. (This will overwrite value from --policy-config)

message_byte_price: Option<Amount>

Set the additional price for each byte in the argument of a user message. (This will overwrite value from --policy-config)

maximum_fuel_per_block: Option<u64>

Set the maximum amount of fuel per block. (This will overwrite value from --policy-config)

maximum_executed_block_size: Option<u64>

Set the maximum size of an executed block. (This will overwrite value from --policy-config)

maximum_bytecode_size: Option<u64>

Set the maximum size of decompressed contract or service bytecode, in bytes. (This will overwrite value from --policy-config)

maximum_blob_size: Option<u64>

Set the maximum size of data blobs, compressed bytecode and other binary blobs, in bytes. (This will overwrite value from --policy-config)

maximum_published_blobs: Option<u64>

Set the maximum number of published blobs per block. (This will overwrite value from --policy-config)

maximum_block_proposal_size: Option<u64>

Set the maximum size of a block proposal, in bytes. (This will overwrite value from --policy-config)

maximum_bytes_read_per_block: Option<u64>

Set the maximum read data per block. (This will overwrite value from --policy-config)

maximum_bytes_written_per_block: Option<u64>

Set the maximum write data per block. (This will overwrite value from --policy-config)

http_allow_list: Option<Vec<String>>

Set the list of hosts that contracts and services can send HTTP requests to.

testing_prng_seed: Option<u64>

Force this wallet to generate keys using a PRNG and a given seed. USE FOR TESTING ONLY.

network_name: Option<String>

A unique name to identify this network.

Watch

Watch the network for notifications.

Fields

chain_id: Option<ChainId>

The chain ID to watch.

raw: bool

Show all notifications from all validators.

Service

Run a GraphQL service to explore and extend the chains of the wallet.

Fields

config: ChainListenerConfig

port: NonZeroU16

The port on which to run the server

Faucet

Run a GraphQL service that exposes a faucet where users can claim tokens. This gives away the chain’s tokens, and is mainly intended for testing.

Fields

chain_id: Option<ChainId>

The chain that gives away its tokens.

port: NonZeroU16

The port on which to run the server

amount: Amount

The number of tokens to send to each new chain.

limit_rate_until: Option<DateTime<Utc>>

The end timestamp: The faucet will rate-limit the token supply so it runs out of money no earlier than this.

config: ChainListenerConfig

Configuration for the faucet chain listener.

PublishModule

Publish module.

Fields

contract: PathBuf

Path to the Wasm file for the application “contract” bytecode.

service: PathBuf

Path to the Wasm file for the application “service” bytecode.

vm_runtime: VmRuntime

The virtual machine runtime to use.

publisher: Option<ChainId>

An optional chain ID to publish the module. The default chain of the wallet is used otherwise.

PublishDataBlob

Publish a data blob of binary data.

Fields

blob_path: PathBuf

Path to data blob file to be published.

publisher: Option<ChainId>

An optional chain ID to publish the blob. The default chain of the wallet is used otherwise.

ReadDataBlob

Verify that a data blob is readable.

Fields

hash: CryptoHash

The hash of the content.

reader: Option<ChainId>

An optional chain ID to verify the blob. The default chain of the wallet is used otherwise.

CreateApplication

Create an application.

Fields

module_id: ModuleId

The module ID of the application to create.

creator: Option<ChainId>

An optional chain ID to host the application. The default chain of the wallet is used otherwise.

json_parameters: Option<String>

The shared parameters as JSON string.

json_parameters_path: Option<PathBuf>

Path to a JSON file containing the shared parameters.

json_argument: Option<String>

The instantiation argument as a JSON string.

json_argument_path: Option<PathBuf>

Path to a JSON file containing the instantiation argument.

required_application_ids: Option<Vec<UserApplicationId>>

The list of required dependencies of application, if any.

PublishAndCreate

Create an application, and publish the required module.

Fields

contract: PathBuf

Path to the Wasm file for the application “contract” bytecode.

service: PathBuf

Path to the Wasm file for the application “service” bytecode.

vm_runtime: VmRuntime

The virtual machine runtime to use.

publisher: Option<ChainId>

An optional chain ID to publish the module. The default chain of the wallet is used otherwise.

json_parameters: Option<String>

The shared parameters as JSON string.

json_parameters_path: Option<PathBuf>

Path to a JSON file containing the shared parameters.

json_argument: Option<String>

The instantiation argument as a JSON string.

json_argument_path: Option<PathBuf>

Path to a JSON file containing the instantiation argument.

required_application_ids: Option<Vec<UserApplicationId>>

The list of required dependencies of application, if any.

Keygen

Create an unassigned key pair.

Assign

Link an owner with a key pair in the wallet to a chain that was created for that owner.

Fields

owner: Owner

The owner to assign.

message_id: MessageId

The ID of the message that created the chain. (This uniquely describes the chain and where it was created.)

RetryPendingBlock

Retry a block we unsuccessfully tried to propose earlier.

As long as a block is pending most other commands will fail, since it is unsafe to propose multiple blocks at the same height.

Fields

chain_id: Option<ChainId>

The chain with the pending block. If not specified, the wallet’s default chain is used.

Wallet(WalletCommand)

Show the contents of the wallet.

Project(ProjectCommand)

Manage Linera projects.

Net(NetCommand)

Manage a local Linera Network.

Storage(DatabaseToolCommand)

Operation on the storage.

HelpMarkdown

Print CLI help in Markdown format, and exit.

ExtractScriptFromMarkdown

Extract a Bash and GraphQL script embedded in a markdown file and print it on stdout.

Fields

path: PathBuf

The source file

pause_after_linera_service: Duration

Insert a pause of N seconds after calls to linera service.

pause_after_gql_mutations: Duration

Insert a pause of N seconds after GraphQL queries.