linera_client
Module persistent

Re-exports

pub use file::File;

pub use memory::Memory;

Modules

file

memory

Traits

LocalPersist

The Persist trait provides a wrapper around a value that can be saved in a persistent way. A minimal implementation provides an Error type, a persist function to persist the value, and an as_mut function to get a mutable reference to the value in memory.

LocalPersistExt

Persist

The Persist trait provides a wrapper around a value that can be saved in a persistent way. A minimal implementation provides an Error type, a persist function to persist the value, and an as_mut function to get a mutable reference to the value in memory.

PersistExt