Enum linera_client::client_options::ClientCommand

pub enum ClientCommand {

    HelpMarkdown,
    Transfer {
        sender: Account,
        recipient: Account,
        amount: Amount,
    },
    OpenChain {
        chain_id: Option<ChainId>,
        public_key: Option<PublicKey>,
        balance: Amount,
    },
    OpenMultiOwnerChain {
        chain_id: Option<ChainId>,
        ownership_config: ChainOwnershipConfig,
        application_permissions_config: ApplicationPermissionsConfig,
        balance: Amount,
    },
    ChangeOwnership {
        chain_id: Option<ChainId>,
        ownership_config: ChainOwnershipConfig,
    },
    ChangeApplicationPermissions {
        chain_id: Option<ChainId>,
        application_permissions_config: ApplicationPermissionsConfig,
    },
    CloseChain {
        chain_id: ChainId,
    },
    LocalBalance {
        account: Option<Account>,
    },
    QueryBalance {
        account: Option<Account>,
    },
    SyncBalance {
        account: Option<Account>,
    },
    Sync {
        chain_id: Option<ChainId>,
    },
    ProcessInbox {
        chain_id: Option<ChainId>,
    },
    QueryValidator {
        address: String,
        chain_id: Option<ChainId>,
        name: Option<ValidatorName>,
    },
    QueryValidators {
        chain_id: Option<ChainId>,
    },
    SetValidator {
        name: ValidatorName,
        address: String,
        votes: u64,
        skip_online_check: bool,
    },
    RemoveValidator {
        name: ValidatorName,
    },
    FinalizeCommittee,
    ResourceControlPolicy {

        block: Option<Amount>,
        fuel_unit: Option<Amount>,
        read_operation: Option<Amount>,
        write_operation: Option<Amount>,
        byte_read: Option<Amount>,
        byte_written: Option<Amount>,
        byte_stored: Option<Amount>,
        operation: Option<Amount>,
        operation_byte: Option<Amount>,
        message: Option<Amount>,
        message_byte: Option<Amount>,
        maximum_fuel_per_block: Option<u64>,
        maximum_executed_block_size: Option<u64>,
        maximum_blob_size: Option<u64>,
        maximum_bytecode_size: Option<u64>,
        maximum_block_proposal_size: Option<u64>,
        maximum_bytes_read_per_block: Option<u64>,
        maximum_bytes_written_per_block: Option<u64>,
    
},
    Benchmark {
        max_in_flight: usize,
        num_chains: usize,
        tokens_per_chain: Amount,
        transactions_per_block: usize,
        fungible_application_id: Option<ApplicationId>,
    },
    CreateGenesisConfig {

        committee_config_path: PathBuf,
        genesis_config_path: PathBuf,
        admin_root: u32,
        initial_funding: Amount,
        start_timestamp: Option<DateTime<Utc>>,
        num_other_initial_chains: u32,
        block_price: Amount,
        fuel_unit_price: Amount,
        read_operation_price: Amount,
        write_operation_price: Amount,
        byte_read_price: Amount,
        byte_written_price: Amount,
        byte_stored_price: Amount,
        operation_price: Amount,
        operation_byte_price: Amount,
        message_price: Amount,
        message_byte_price: Amount,
        maximum_fuel_per_block: Option<u64>,
        maximum_executed_block_size: Option<u64>,
        maximum_bytecode_size: Option<u64>,
        maximum_blob_size: Option<u64>,
        maximum_block_proposal_size: Option<u64>,
        maximum_bytes_read_per_block: Option<u64>,
        maximum_bytes_written_per_block: Option<u64>,
        testing_prng_seed: Option<u64>,
        network_name: Option<String>,
    
},
    Watch {
        chain_id: Option<ChainId>,
        raw: bool,
    },
    Service {
        config: ChainListenerConfig,
        port: NonZeroU16,
    },
    Faucet {
        chain_id: Option<ChainId>,
        port: NonZeroU16,
        amount: Amount,
        limit_rate_until: Option<DateTime<Utc>>,
        config: ChainListenerConfig,
    },
    PublishBytecode {
        contract: PathBuf,
        service: PathBuf,
        publisher: Option<ChainId>,
    },
    PublishDataBlob {
        blob_path: PathBuf,
        publisher: Option<ChainId>,
    },
    ReadDataBlob {
        hash: CryptoHash,
        reader: Option<ChainId>,
    },
    CreateApplication {
        bytecode_id: BytecodeId,
        creator: Option<ChainId>,
        json_parameters: Option<String>,
        json_parameters_path: Option<PathBuf>,
        json_argument: Option<String>,
        json_argument_path: Option<PathBuf>,
        required_application_ids: Option<Vec<UserApplicationId>>,
    },
    PublishAndCreate {
        contract: PathBuf,
        service: PathBuf,
        publisher: Option<ChainId>,
        json_parameters: Option<String>,
        json_parameters_path: Option<PathBuf>,
        json_argument: Option<String>,
        json_argument_path: Option<PathBuf>,
        required_application_ids: Option<Vec<UserApplicationId>>,
    },
    RequestApplication {
        application_id: UserApplicationId,
        target_chain_id: Option<ChainId>,
        requester_chain_id: Option<ChainId>,
    },
    Keygen,
    Assign {
        key: PublicKey,
        message_id: MessageId,
    },
    RetryPendingBlock {
        chain_id: Option<ChainId>,
    },
    Wallet(WalletCommand),
    Project(ProjectCommand),
    Net(NetCommand),
    Storage(DatabaseToolCommand),

}

Variants

HelpMarkdown

Print CLI help in Markdown format, and exit.

Transfer

Transfer funds

Fields

sender: Account

Sending chain ID (must be one of our chains)

recipient: Account

Recipient account

amount: Amount

Amount to transfer

OpenChain

Open (i.e. activate) a new chain deriving the UID from an existing one.

Fields

chain_id: Option<ChainId>

Chain ID (must be one of our chains).

public_key: Option<PublicKey>

Public key of the new owner (otherwise create a key pair and remember it)

balance: Amount

The initial balance of the new chain. This is subtracted from the parent chain’s balance.

OpenMultiOwnerChain

Open (i.e. activate) a new multi-owner chain deriving the UID from an existing one.

Fields

chain_id: Option<ChainId>

Chain ID (must be one of our chains).

ownership_config: ChainOwnershipConfig

application_permissions_config: ApplicationPermissionsConfig

balance: Amount

The initial balance of the new chain. This is subtracted from the parent chain’s balance.

ChangeOwnership

Change who owns the chain, and how the owners work together proposing blocks.

Specify the complete set of new owners, by public key. Existing owners that are not included will be removed.

Fields

chain_id: Option<ChainId>

The ID of the chain whose owners will be changed.

ownership_config: ChainOwnershipConfig

ChangeApplicationPermissions

Changes the application permissions configuration.

Fields

chain_id: Option<ChainId>

The ID of the chain to which the new permissions will be applied.

application_permissions_config: ApplicationPermissionsConfig

CloseChain

Close an existing chain.

A closed chain cannot execute operations or accept messages anymore. It can still reject incoming messages, so they bounce back to the sender.

Fields

chain_id: ChainId

Chain ID (must be one of our chains)

LocalBalance

Read the current native-token balance of the given account directly from the local state.

NOTE: The local balance does not reflect messages that are waiting to be picked in the local inbox, or that have not been synchronized from validators yet. Use linera sync then either linera query-balance or linera process-inbox && linera local-balance for a consolidated balance.

Fields

account: Option<Account>

The account to read, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

QueryBalance

Simulate the execution of one block made of pending messages from the local inbox, then read the native-token balance of the account from the local state.

NOTE: The balance does not reflect messages that have not been synchronized from validators yet. Call linera sync first to do so.

Fields

account: Option<Account>

The account to query, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

SyncBalance

(DEPRECATED) Synchronize the local state of the chain with a quorum validators, then query the local balance.

This command is deprecated. Use linera sync && linera query-balance instead.

Fields

account: Option<Account>

The account to query, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

Sync

Synchronize the local state of the chain with a quorum validators.

Fields

chain_id: Option<ChainId>

The chain to synchronize with validators. If omitted, synchronizes the default chain of the wallet.

ProcessInbox

Process all pending incoming messages from the inbox of the given chain by creating as many blocks as needed to execute all (non-failing) messages. Failing messages will be marked as rejected and may bounce to their sender depending on their configuration.

Fields

chain_id: Option<ChainId>

The chain to process. If omitted, uses the default chain of the wallet.

QueryValidator

Show the version and genesis config hash of a new validator, and print a warning if it is incompatible. Also print some information about the given chain while we are at it.

Fields

address: String

The new validator’s address.

chain_id: Option<ChainId>

The chain to query. If omitted, query the default chain of the wallet.

name: Option<ValidatorName>

The public key of the validator. If given, the signature of the chain query info will be checked.

QueryValidators

Show the current set of validators for a chain. Also print some information about the given chain while we are at it.

Fields

chain_id: Option<ChainId>

The chain to query. If omitted, query the default chain of the wallet.

SetValidator

Add or modify a validator (admin only)

Fields

name: ValidatorName

The public key of the validator.

address: String

Network address

votes: u64

Voting power

skip_online_check: bool

Skip the version and genesis config checks.

RemoveValidator

Remove a validator (admin only)

Fields

name: ValidatorName

The public key of the validator.

FinalizeCommittee

Deprecates all committees except the last one.

ResourceControlPolicy

View or update the resource control policy

Fields

block: Option<Amount>

Set the base price for creating a block.

fuel_unit: Option<Amount>

Set the price per unit of fuel.

read_operation: Option<Amount>

Set the price per read operation.

write_operation: Option<Amount>

Set the price per write operation.

byte_read: Option<Amount>

Set the price per byte read.

byte_written: Option<Amount>

Set the price per byte written.

byte_stored: Option<Amount>

Set the price per byte stored.

operation: Option<Amount>

Set the base price of sending an operation from a block..

operation_byte: Option<Amount>

Set the additional price for each byte in the argument of a user operation.

message: Option<Amount>

Set the base price of sending a message from a block..

message_byte: Option<Amount>

Set the additional price for each byte in the argument of a user message.

maximum_fuel_per_block: Option<u64>

Set the maximum amount of fuel per block.

maximum_executed_block_size: Option<u64>

Set the maximum size of an executed block, in bytes.

maximum_blob_size: Option<u64>

Set the maximum size of data blobs, compressed bytecode and other binary blobs, in bytes.

maximum_bytecode_size: Option<u64>

Set the maximum size of decompressed contract or service bytecode, in bytes.

maximum_block_proposal_size: Option<u64>

Set the maximum size of a block proposal, in bytes.

maximum_bytes_read_per_block: Option<u64>

Set the maximum read data per block.

maximum_bytes_written_per_block: Option<u64>

Set the maximum write data per block.

Benchmark

Send one transfer per chain in bulk mode

Fields

max_in_flight: usize

Maximum number of blocks in flight

num_chains: usize

How many chains to use for the benchmark

tokens_per_chain: Amount

How many tokens to assign to each newly created chain. These need to cover the transaction fees per chain for the benchmark.

transactions_per_block: usize

How many transactions to put in each block.

fungible_application_id: Option<ApplicationId>

The application ID of a fungible token on the wallet’s default chain. If none is specified, the benchmark uses the native token.

CreateGenesisConfig

Create genesis configuration for a Linera deployment. Create initial user chains and print information to be used for initialization of validator setup. This will also create an initial wallet for the owner of the initial “root” chains.

Fields

committee_config_path: PathBuf

Sets the file describing the public configurations of all validators

genesis_config_path: PathBuf

The output config path to be consumed by the server

admin_root: u32

Index of the admin chain in the genesis config

initial_funding: Amount

Known initial balance of the chain

start_timestamp: Option<DateTime<Utc>>

The start timestamp: no blocks can be created before this time.

num_other_initial_chains: u32

Number of initial (aka “root”) chains to create in addition to the admin chain.

block_price: Amount

Set the base price for creating a block.

fuel_unit_price: Amount

Set the price per unit of fuel.

read_operation_price: Amount

Set the price per read operation.

write_operation_price: Amount

Set the price per write operation.

byte_read_price: Amount

Set the price per byte read.

byte_written_price: Amount

Set the price per byte written.

byte_stored_price: Amount

Set the price per byte stored.

operation_price: Amount

Set the base price of sending an operation from a block..

operation_byte_price: Amount

Set the additional price for each byte in the argument of a user operation.

message_price: Amount

Set the base price of sending a message from a block..

message_byte_price: Amount

Set the additional price for each byte in the argument of a user message.

maximum_fuel_per_block: Option<u64>

Set the maximum amount of fuel per block.

maximum_executed_block_size: Option<u64>

Set the maximum size of an executed block.

maximum_bytecode_size: Option<u64>

Set the maximum size of decompressed contract or service bytecode, in bytes.

maximum_blob_size: Option<u64>

Set the maximum size of data blobs, compressed bytecode and other binary blobs, in bytes.

maximum_block_proposal_size: Option<u64>

Set the maximum size of a block proposal, in bytes.

maximum_bytes_read_per_block: Option<u64>

Set the maximum read data per block.

maximum_bytes_written_per_block: Option<u64>

Set the maximum write data per block.

testing_prng_seed: Option<u64>

Force this wallet to generate keys using a PRNG and a given seed. USE FOR TESTING ONLY.

network_name: Option<String>

A unique name to identify this network.

Watch

Watch the network for notifications.

Fields

chain_id: Option<ChainId>

The chain ID to watch.

raw: bool

Show all notifications from all validators.

Service

Run a GraphQL service to explore and extend the chains of the wallet.

Fields

config: ChainListenerConfig

port: NonZeroU16

The port on which to run the server

Faucet

Run a GraphQL service that exposes a faucet where users can claim tokens. This gives away the chain’s tokens, and is mainly intended for testing.

Fields

chain_id: Option<ChainId>

The chain that gives away its tokens.

port: NonZeroU16

The port on which to run the server

amount: Amount

The number of tokens to send to each new chain.

limit_rate_until: Option<DateTime<Utc>>

The end timestamp: The faucet will rate-limit the token supply so it runs out of money no earlier than this.

config: ChainListenerConfig

Configuration for the faucet chain listener.

PublishBytecode

Publish bytecode.

Fields

contract: PathBuf

Path to the Wasm file for the application “contract” bytecode.

service: PathBuf

Path to the Wasm file for the application “service” bytecode.

publisher: Option<ChainId>

An optional chain ID to publish the bytecode. The default chain of the wallet is used otherwise.

PublishDataBlob

Publish a data blob of binary data.

Fields

blob_path: PathBuf

Path to data blob file to be published.

publisher: Option<ChainId>

An optional chain ID to publish the blob. The default chain of the wallet is used otherwise.

ReadDataBlob

Verify that a data blob is readable.

Fields

hash: CryptoHash

The hash of the content.

reader: Option<ChainId>

An optional chain ID to verify the blob. The default chain of the wallet is used otherwise.

CreateApplication

Create an application.

Fields

bytecode_id: BytecodeId

The bytecode ID of the application to create.

creator: Option<ChainId>

An optional chain ID to host the application. The default chain of the wallet is used otherwise.

json_parameters: Option<String>

The shared parameters as JSON string.

json_parameters_path: Option<PathBuf>

Path to a JSON file containing the shared parameters.

json_argument: Option<String>

The instantiation argument as a JSON string.

json_argument_path: Option<PathBuf>

Path to a JSON file containing the instantiation argument.

required_application_ids: Option<Vec<UserApplicationId>>

The list of required dependencies of application, if any.

PublishAndCreate

Create an application, and publish the required bytecode.

Fields

contract: PathBuf

Path to the Wasm file for the application “contract” bytecode.

service: PathBuf

Path to the Wasm file for the application “service” bytecode.

publisher: Option<ChainId>

An optional chain ID to publish the bytecode. The default chain of the wallet is used otherwise.

json_parameters: Option<String>

The shared parameters as JSON string.

json_parameters_path: Option<PathBuf>

Path to a JSON file containing the shared parameters.

json_argument: Option<String>

The instantiation argument as a JSON string.

json_argument_path: Option<PathBuf>

Path to a JSON file containing the instantiation argument.

required_application_ids: Option<Vec<UserApplicationId>>

The list of required dependencies of application, if any.

RequestApplication

Request an application from another chain, so it can be used on this one.

Fields

application_id: UserApplicationId

The ID of the application to request.

target_chain_id: Option<ChainId>

The target chain on which the application is already registered. If not specified, the chain on which the application was created is used.

requester_chain_id: Option<ChainId>

The owned chain on which the application is missing.

Keygen

Create an unassigned key-pair.

Assign

Link a key owned by the wallet to a chain that was just created for that key.

Fields

key: PublicKey

The public key to assign.

message_id: MessageId

The ID of the message that created the chain. (This uniquely describes the chain and where it was created.)

RetryPendingBlock

Retry a block we unsuccessfully tried to propose earlier.

As long as a block is pending most other commands will fail, since it is unsafe to propose multiple blocks at the same height.

Fields

chain_id: Option<ChainId>

The chain with the pending block. If not specified, the wallet’s default chain is used.

Wallet(WalletCommand)

Show the contents of the wallet.

Project(ProjectCommand)

Manage Linera projects.

Net(NetCommand)

Manage a local Linera Network.

Storage(DatabaseToolCommand)

Operation on the storage.